requirements.txt
SQLAlchemy
Flask
Flask-SQLAlchemy
Flask-Security
Flask-WTF
wtforms
python-dotenv
mariadb
email_validator
bcrypt